requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
Pillow
uwsgi>=2.0.20,<2.1
playwright
crawl4ai
langchain-core
langchain
langchain-openai
langgraph
chromadb
pypdf
python-dotenv
openai
langgraph-sdk<0.4.0,>=0.3.0
ddtrace

